+++ package.json
@@ -4,36 +4,42 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve sandbox",
-    "build": "ng build ui",
-    "postbuild": "node ./bin/postbuild.js && npm run package && npm run install:library",
-    "package": "tar -cvf ./dist/ui.tar.gz -C dist/ui .",
-    "postinstall": "npm run build && npm i --save ./dist/ui.tar.gz",
-    "install:library": "npm i --save ./dist/ui.tar.gz"
+    "build": "npm run build:core && npm run build:forms",
+    "build:core": "ng build core",
+    "build:forms": "ng build forms",
+    "install-library": "npm run install-library:core && npm run install-library:forms",
+    "install-library:core": "npm i --save ./dist/core.tar.gz",
+    "install-library:forms": "npm i --save ./dist/forms.tar.gz",
+    "postbuild": "node ./bin/postbuild.js && npm run package && npm run install-library",
+    "package": "npm run package:core && npm run package:forms",
+    "package:core": "tar -cvf ./dist/core.tar.gz -C dist/core .",
+    "package:forms": "tar -cvf ./dist/forms.tar.gz -C dist/forms .",
+    "postinstall": "npm run build"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.0.6",
-    "@angular/cdk": "^11.0.4",
-    "@angular/common": "~11.0.6",
-    "@angular/compiler": "~11.0.6",
-    "@angular/core": "~11.0.6",
-    "@angular/forms": "~11.0.6",
-    "@angular/platform-browser": "~11.0.6",
-    "@angular/platform-browser-dynamic": "~11.0.6",
-    "@angular/router": "~11.0.6",
-    "@tsmonkeypatch/common": "file:dist/common.tar.gz",
-    "@tsmonkeypatch/ui": "file:dist/ui.tar.gz",
-    "bootstrap-scss": "^4.5.3",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.2"
+    "@angular/animations": "11.0.6",
+    "@angular/cdk": "11.0.4",
+    "@angular/common": "11.0.6",
+    "@angular/compiler": "11.0.6",
+    "@angular/core": "11.0.6",
+    "@angular/forms": "11.0.6",
+    "@angular/platform-browser": "11.0.6",
+    "@angular/platform-browser-dynamic": "11.0.6",
+    "@angular/router": "11.0.6",
+    "@tsmonkeypatch/core": "file:dist/core.tar.gz",
+    "@tsmonkeypatch/forms": "file:dist/forms.tar.gz",
+    "bootstrap-scss": "4.5.3",
+    "rxjs": "6.6.0",
+    "tslib": "2.0.0",
+    "zone.js": "0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1100.6",
-    "@angular/cli": "~11.0.6",
-    "@angular/compiler-cli": "~11.0.6",
-    "@types/node": "^12.11.1",
-    "ng-packagr": "^11.0.0",
-    "typescript": "~4.0.2"
+    "@angular-devkit/build-angular": "0.1100.6",
+    "@angular/cli": "11.0.6",
+    "@angular/compiler-cli": "11.0.6",
+    "@types/node": "12.11.1",
+    "ng-packagr": "11.0.0",
+    "typescript": "4.0.2"
   }
 }
