+++ projects/core/README.md
