+++ projects/core/common/package.json
