+++ projects/core/common/src/directives/sticky-directive.ts
