+++ projects/core/karma.conf.js
