+++ projects/core/ng-package.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/core",
+  "lib": {
+    "entryFile": "public-api.ts",
+    "umdModuleIds": {
+      "@tsmonkeypatch/core": "@tsmonkeypatch/core"
+    }
+  }
+}
