+++ projects/core/overlay/package.json
@@ -3,7 +3,10 @@
   "ngPackage": {
     "dest": "../../dist/ui",
     "lib": {
-        "entryFile": "public-api.ts"
+        "entryFile": "public-api.ts",
+        "umdModuleIds": {
+          "tsmonkeypatch-overlay": "TsMonkeyPatchOverlay"
+        }
     }
   }
 }
