+++ projects/core/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@tsmonkeypatch/core",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "11.0.6",
+    "@angular/core": "11.0.6"
+  },
+  "dependencies": {
+    "tslib": "2.0.0"
+  }
+}
