+++ projects/core/stepper/package.json
