+++ projects/core/tsconfig.lib.json
