+++ projects/core/tsconfig.lib.prod.json
