--- projects/ui/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "@tsmonkeypatch/ui",
-  "version": "0.0.1",
-  "peerDependencies": {
-    "@angular/common": "^11.0.6",
-    "@angular/core": "^11.0.6"
-  },
-  "dependencies": {
-    "tslib": "^2.0.0"
-  }
-}
